requirements.txt
et-xmlfile==1.1.0
greenlet==1.1.2
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.0
psycopg2-binary==2.9.3
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
xlrd==2.0.1
XlsxWriter==3.0.2
